Art - misc.
- Checkmarks: these should appear over the items that have been
enqueued. Orange for the chef and purple for Batilda.
- Money.png: the first sprite (single coin) is for particle effects. The
other 3 are small, medium, and large payments that appear on the counter
after the customer has left. The particle effect should appear once
Batilda comes by to pick up the money (a burst of coins that is either
small, medium, or large depending on the payment)